# static/swagger.yaml
openapi: 3.0.3
info:
  title: API Домашней Библиотеки
  description: REST API для управления домашней библиотекой
  version: 1.0.0
paths:
  /books:
    get:
      tags:
        - Book
      summary: Получить все книги
      responses:
        '200':
          description: Успешное выполнение
        '404':
          description: Книги не найдены
    post:
      tags:
        - Book
      summary: Добавить новую книгу
      requestBody:
        description: Данные книги
        required: true
        content:
          application/json:
            example:
              title: "Новая Книга"
              author: "Автор"
              genre: "Жанр"
              year: "Год выхода"
      responses:
        '201':
          description: Книга успешно добавлена
          content:
            application/json:
              example:
               title: "Новая Книга"
               author: "Автор"
               genre: "Жанр"
               year: "Год выхода"
        '400':
          description: Неверные данные
  /books/{IdBooks}:
    get:
      tags:
        - Book
      summary: Получить книгу по идентификатору
      parameters:
        - name: IdBooks
          in: path
          required: true
          description: Идентификатор книги
          schema:
            type: integer
      responses:
        '200':
          description: Успешное выполнение
        '404':
          description: Книга не найдена
    put:
      tags:
        - Book
      summary: Обновить книгу по идентификатору
      parameters:
        - name: IdBooks
          in: path
          required: true
          description: Идентификатор книги
          schema:
            type: integer
      requestBody:
        description: Обновленные данные книги
        required: true
        content:
          application/json:
            example:
              title: "Новая Книга"
              author: "Автор"
              genre: "Жанр"
              year: "Год выхода"
      responses:
        '200':
          description: Книга успешно обновлена
          content:
            application/json:
              example:
                id: "5"
                title: "Новая Книга"
                author: "Автор"
                genre: "Жанр"
                year: "Год выхода"
        '400':
          description: Неверные данные
        '404':
          description: Книга не найдена
    delete:
      tags:
        - Book
      summary: Удалить книгу по идентификатору
      parameters:
        - name: IdBooks
          in: path
          required: true
          description: Идентификатор книги
          schema:
            type: integer
      responses:
        '200':
          description: Книга успешно удалена
        '404':
          description: Книга не найдена